Refactor Kotlin type-awareness flow and tighten Kotlin checks
